fix(v1.7): add missing auth/scopes module extracted from server.ts
The scope constants and resolveOAuthScopes() were extracted into a
shared module but the new file was never committed, breaking the
CI build.
